app/requirements.txt
pandas>=1.5.0
plotly>=5.9.0
psycopg2-binary==2.9.3
matplotlib==3.5.1
flask==2.0.3
werkzeug==2.0.3
